backend_DRF/pyproject.toml
[tool.poetry]
name = "backend_drf"
version = "0.1.0"
description = ""
authors = ["Antoine Dorelon <antoinedorelon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
pytest = "^6.2.4"
pytest-django = "^4.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
